+++ sample/src/main/java/com/timmytruong/calendar/ui/screen/util/PopUps.kt
@@ -1,0 +1,10 @@
+package com.timmytruong.calendar.ui.screen.util
+
+import android.content.Context
+import android.widget.Toast
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
+
+internal fun LocalDate.showToast(context: Context) {
+    Toast.makeText(context, format(DateTimeFormatter.ISO_LOCAL_DATE), Toast.LENGTH_SHORT).show()
+}
